Thread through enclosingMethod from SendResponse
The only place that issues a SendResponse for an LSPQuery is
processBinding, which means we only issue a SendResponse when we know
the enclosing method where that send lives.

Threading that information through the response makes it way easier to
implement completion for local variables, because now I can ask "please
list the locals for this method" rather than having to first find the
method whose locals I need to list.