Fix CSV << return type
I ran into this (apparent) bug when updating our codebase to the latest Sorbet. In a few places we chain usages of `<<` for example:

```
csv << row << [] 
```
And this works fine because `<<` returns a `CSV`. This is easy enough for us to work around but I think this signature is more correct.