Add try_deserialize which returns nil (#2096)
Sometimes it'd be nice to return `nil` instead of `KeyError` when
deserialization fails. `deserialize` is a part of T::Props::CustomType,
so we can't change the API for that one. Instead, we should introduce a
new method. I've chosen the name `try_deserialize`.

At least two people have asked for this on Slack, and it was a pretty
obvious gap when writing the documentation, so considering it's easy to
implement, I think we should just do it.