Implement completion for local variables (#2106)
* Thread through enclosingMethod from SendResponse

The only place that issues a SendResponse for an LSPQuery is
processBinding, which means we only issue a SendResponse when we know
the enclosing method where that send lives.

Threading that information through the response makes it way easier to
implement completion for local variables, because now I can ask "please
list the locals for this method" rather than having to first find the
method whose locals I need to list.

* Implement LSPTypechecker::localsForMethod

This adds a new public API to LSPTypechecker which returns a vector of
LocalVariables given a method symbol.

Internally, it duplicates a lot of code with LSPTypechecker::query, but
I think that's fine because it makes it easy to see what's going on.

The basic algorithm is:

- Use the method symbol's locs to get all the files that might have
  defined this method.
- Run a TreeMap over those those files to collect local variables.
- The TreeMap visits MethodDefs and Assigns and checks to see if we
  should accumulate that node's locals (i.e., if we're currently inside
  the the target method).

* Implement completion for local variables

Uses SendResponse::enclosingMethod + LSPTypechecker::localsForMethod to
show completion items for local variables. The implementation itself is
very similar to how completion for keywords works.

* Add lots of tests

* Add another test

* Also support static init

* getIndex -> getIndexed

No one is actually calling this method, and I like it to be in the past
tense because I plan to make a new function called getResolved.

* Move localsForMethod out of LSPTypechecker

Instead, we expose the ability to ask for the resolved tree from
LSPTypechecker, which is a more general interface.