Shorten & customize the detail field for keywords
VS Code shows the `detail` as a monospace string (unconditionally: there
is no way to set the MarkupKind for the `detail`). This wraps in a
really weird way. This PR shortens the detail for most keywords, and
puts the documentation into the expanded view (which is better anyways,
because no one was going to be able to read the full documentation for
some of these items anyways).

In addition to shortening, I took the chance to add custom details for
some keywords:

- Sometimes, the detail makes it more clear that it'll insert a snippet.
- Sometimes, the detail makes it more clear what the snippet will insert
  (for example, `if/end` makes it clear that it won't add an `else`).

Before:

TODO image

After:

TODO image