Show sig + def in completion results
For hover, we show both the sig and method def for a method, because
some information like keyword / defaulted args are only present on the
method def.

We were not showing this in completion.

I made a new helper called prettyTypeForMethod which always calls the
`sig` part and the `def` part, and removed the two individual components
from the header to force everyone to go through the other two methods.

After this change, we show the `def` for completion results too.