Shorten detail, format documentation as markdown
For TypeScript, the `detail` field shows the type.
But for Sorbet, the type is frequently a sig and is huge (fully
qualified class names, lots of args, a sig and a def, ...). They get
ugly really quickly.

I've opted to put the sig + def only in the documentation field, so that
people can see it when they press `^Space` (VS Code already hides the
details for completion items that are not active). Unlike the detail,
LSP allows for the documentation to declare its MarkupKind (so we can
have markdown there).

Then the question becomes: what should we put in the detail?

In the private beta, someone had asked for a way to see which class
owned this method, so it seemed like a way to solve two birds with one
stone.