Remove localsForMethod from LSPLoop (#2146)
This can be a helper function instead of a method.

If we end up wanting to call it from more than one place, we can put it
somewhere publically accessible. (LSPLoop feels like a grab bag of lots
of unrelated things, so in my mind the fewer things there, the better).