Relax bar of "useful guess" for SUGGEST_SIG query
For the purpose of autocorrects / code actions, we only insert a sig
when we guessed a "useful" sig, namely at least one non-T.untyped
somewhere (this was partially motivated by the effort to let multiple
runs of sig suggestion incrementally improve the sigs each time).

But for the purpose of autocomplete, even so much as having names for
all the arguments already typed out is a benefit, regardless of whether
they're all untyped.