Replace T.untyped in sig snippet with tabstops
This lets people cycle through the `T.untyped` as if they're holes to be
filled in.

The code is kind of ugly but I couldn't think of a much easier way to
implement it without making `SigSuggestion.cc` aware of whether the LSP
client supports snippets (which seems like it violates the concerns of
what `infer/` should have to know about).