Allow srb to check for sorbet on PATH
For those that are using JRuby -- unfortunately, sorbet-static can never
be installed!

sorbet-static is platform dependent and has the gem name such as:
`sorbet-static-0.4.5055-universal-darwin-19.gem`
(darwin being OSX & universal the CPU type)

However when running on JRuby platform the only platforms available are:
```bash
$ gem env
...
  - RUBYGEMS PLATFORMS:
    - ruby
    - universal-java-1.8
```

Whereas on MRI it will list
```bash
$ gem env
  - RUBYGEMS PLATFORMS:
    - ruby
    - universal-darwin-18
```

Given this dilemma, how can we use sorbet-static on JRuby ?
I propose that we leverage platform package managers such as: homebrew
or yum etc.. to deliver sorbet. The `srb` script is changed to check if
sorbet is present on the path.