Automatically insert snippet for required blocks
When we can tell from the types that a method must take a block, we can
automatically insert the block snippet.

This snippet unconditionally makes the block snippet multi-line.

Sometimes people might want / need a single line snippet. I suggest that
we merge this as is and if people ask for single line snippets, that we
implement that as a refactoring / code action, or maybe even just revert
this.