Fix failing tests
It looks like the snippet isn't properly indented, but VS Code actually
indents it the right amount depending on context.

I'm not sure whether re-indenting the snippet is part of the spec or
not; we might want to make the test driver attempt to do something
similar, but for now, it's enough for me to see the changes this way.