Add static error for instantiating an abstract class
The runtime has a check for this but it can be defeated if the abstract
class defines `initialize`. Efficently fixing that problem in the
runtime involves using `prepend`, which is risky as it invalidates
assumptions scattered around the codebase.

This only checks calling `new` on a class marked directly with
`abstract!`. Subclasses of abstract classes that are not concrete
already get errors about missing definitions so we should be okay there.

Thanks to @jez for suggesting where to put this check!

Closes #2143