Automatically insert snippet for required blocks (#2219)
* Automatically insert snippet for required blocks

When we can tell from the types that a method must take a block, we can
automatically insert the block snippet.

This snippet unconditionally makes the block snippet multi-line.

Sometimes people might want / need a single line snippet. I suggest that
we merge this as is and if people ask for single line snippets, that we
implement that as a refactoring / code action, or maybe even just revert
this.

* Fix failing tests

It looks like the snippet isn't properly indented, but VS Code actually
indents it the right amount depending on context.

I'm not sure whether re-indenting the snippet is part of the spec or
not; we might want to make the test driver attempt to do something
similar, but for now, it's enough for me to see the changes this way.