Add failing (flaky) test
One of `sig_redefinition__1.rb` or `sig_redefinition__2.rb` would get
the primary loc for either `foo` or `bar` depending on at random which
file was processed first.

Depending on which of the two files became the primary loc, only in that
file would sig suggestion trigger. I noticed this because when I was
testing LSP locally I frequently use

```ruby
class A
  def foo; end
end
```

in small test files. When multiple such test files are in a folder,
Sorbet sees all the `A#foo` method defs as (compatible) redefinitions of
each other, and accumulates all their Locs onto one Symbol.