Remove owner from QueryResponse's
IdentResponse, LiteralResponse, and ConstantResponse all have an `owner`
field on them which doesn't really make sense.

- It's completely unused for ConstantResponse and LiteralResponse.
- For IdentResponse, it's used, but it's kind of strange: when the
  IdentResponse is for an instance variable (which would be a Symbol
  which is "owned" by a class), the `owner` would actually be the method
  that that instance variable was used in (i.e., the lexical scope).

Add into the mix that I want to actually track `owner` on
ConstantResponse in the Symbol sense of the word, and I figure that's
enough to get rid of this.

For the one place where we need it (IdentResponse), I'm going to
re-introduce something like `SymbolRef lexicalScope` in the next commit.