Introduce release file for java platform
After haven spoken to @DarkDimius over slack regarding #2213
we discussed the option of creating a java platform release that
contains both linux & mac static binaries --
srb would be augmented to select the appropriate binary given the
`uname` of the system if the platform is java.

This a submission for a ruby script that will pull down pre-released
sorbet-static gems for linux & darwin and create a java one.