Don't put repeated args in the completion snippet
Repeated args aren't required (it's possible to type check without
providing a single one) so we shouldn't require that it be filled out
in the snippet.

For `**x` args in particular it was super annoying, because this shows
up as both `isKeyword` and `isRepeated` so we'd even put an `x: ${1}` in
the snippet, when in fact, that `x` isn't the same as the `x` in `**x`.