Test apply-completion for constants
The test harness isn't a complete LSP client... It only supports making
edits that use the textEdit property of a CompletionItem. In order to
test `apply-completion`, I needed to change getCompletionItemForConstant
to craft a TextEdit object the same way we do for all other kinds of
completions.

There are good reasons for preferring TextEdit (the default is to use
the CompletionItem label, and VS Code takes a liberty with what edit to
perform in the document). TextEdit gives us complete control.