Fix publish-ruby-gems.sh (#2305)
This is a fix ontop of #2298 which changed slightly `publish-ruby-gems.sh`
The Darwin platform is the only gem with multiple files, so changed the
publishing code to be explicit.

The benefit of this approach to the previous glob file approach, is that
the version can be checked for each platform number (i.e.
universal-darwin-18) if it has been published before.