Fix suggest-sig-override-edge test
At some point, we changed this file to stop testing anything useful.
(My guess is around the time that we removed `implementation` we edited
this test incorrectly.)

The original point of this test was that we needed some way to insert
`generated` up into the parent's sig when suggesting a sig on a child.

We never need to do that anymore because of the changes to `override`
let us simplify some of that logic (so that as long as the parent method
exists at all, it's ok to put `override` on the child).

But as a side effect, it meant that we stopped testing the ordering of
the builders. So this test now tests the order of the builders (which is
wrong; fixed in the next commit).