Fix crash in parser for heredocs
In the included test case, Dedenter::dedent starts with `str="\tZ"` and
`spacesToRemove=1`.

The problem is that the case for handling tabs tries to decrement
`spacesToRemove` by 8, which causes unsigned integer underflow, so the
loop guard fails to trip, and we start processing non-whitespace
characters (`Z`) and try to treat them as whitespace to be removed.

The fix is to never underflow the unsigned int, by handling the case
when it would go negative to early exit.