Switch back to gem list
The output of gem fetch was proving to be too hard to work with (it
raised an exit code when the fetch failed, which would have meant
needing to silence the exit code of the `gem fetch` to not cause the
whole pipe to fail, which makes the overall script more brittle because
it could fail for a reason that's not "the gem doesn't exist" but
something else like "the network is down").

Switching back to gem list is also about 15 seconds faster.