Remove .clang-tidy file
We don't enforce clang-tidy in CI, which means there are lots of
clang-tidy errors in our codebase.

Also, clangd uses the presence of a `.clang-tidy` file at the top level
as an indiciation that it should implicitly assume `clangd --clang-tidy`
and report clang-tidy violations as errors in the editor.

I can't see a way to disable this via the CLI (i.e., there's no CLI
override like `--no-clang-tidy`). With that in mind, I'd like to remove
the file.

I'd also be in favor of adding it back with a list of exclusions or a
diff full of fixes, plus checking it in CI (though I think the reason
why we weren't checking in CI was because it took longer than it did to
compile).

I'm open to a discussion here.