Close issue about incorrect docs for constants
The incorrect behavior had to do with the handling of type members. I
thought we were wrong at first. I don't know why I thought that, because
the test passed in CI that I thought would fail, and then I confirmed
that it also worked in the editor manually.

The only thing left that could be construed as "wrong" is that for class
aliases we show the docs for the dealiased symbol, which I think is
actually preferable in some cases, so I'm going to close this.