Sprinkle some dealiases in constant completion
We need to use the source symbol (`maybeAlias`) for anything where the
name of the symbol as typed by the user is important, but the target
symbol (`what`) for anything where the types / arity are important.

In particular, this change should only manifest as a difference in
behavior for what's shown in the `documentation` field, which is
currrently untested (it's tested indirectly by sharing a large portion
of code with hover).