Sprinkle some dealiases in constant completion (#2436)
* Sprinkle some dealiases in constant completion

We need to use the source symbol (`maybeAlias`) for anything where the
name of the symbol as typed by the user is important, but the target
symbol (`what`) for anything where the types / arity are important.

In particular, this change should only manifest as a difference in
behavior for what's shown in the `documentation` field, which is
currrently untested (it's tested indirectly by sharing a large portion
of code with hover).

* Consolidate constant type handling in prettyTypeForConstant

Turns out, resultType was unused in getCompletionItemForConstant, and
instead it was exclusively passing its `type` argument forwards to
prettyTypeForConstant. It would have been nice to get a dead code
warning there...

By putting all this logic in prettyTypeForConstant, it meanns that it
will be shared with hover.