Delete dead code from resolver.cc
I'm pretty sure this code is dead. Deleting it doesn't make a single
test fail. I'm going to test on pay-server to be sure, but I think we
can safely delete this unless someone else knows better.

This makes a refactor that I have to do for constant completion a little
bit easier.