Add special casing in runtime for writer methods
This commit tries to identify writer methods generated by attr_writer or
attr_accessor methods properly so that type-checking can be performed at
runtime without errors.

Since writer methods have a single parameter with a declared name,
currently the runtime fails matching the parameter to the one declared
in the signature.

The fix uses heuristics to try to identify which methods are writer
methods with a supplied signature and modifies the parameter list so
that runtime type-checking can be performed.

More details on the strategy can be found here: https://github.com/sorbet/sorbet/issues/1239#issuecomment-577382200