Swap this_file & decl_file in sealed_violation_whitelist
The whole point of this regex was to be able to opt entire swaths of a
project (like, various folders) into skipping the sealed definition
checks. For that purpose, it was implemented backwards... whoops.

The bug was masked by the fact that I also swapped the regex in the test
I wrote.