Support rbenv generated gem paths (#2597)
Update the regular expression in TracepointSerializer to account for
a gem path created by rbenv. Specifically, we should support the
following path:

`/mnt/rbenv/versions/2.3.1/gems/gems/mygem-0.2/lib/mygem.rb`

Note the `/gems/gems/` in the middle of this path where the old
regular expression assumes a ruby version (but rbenv puts the ruby
version at the beginning of the folder structure).