But actually, this can be synthesized
I tested this on pay-server, and this one which was not being marked as
RewriterSynthesized is actually safe to mark as such, and probably
should be for consistency. That means that the call site in Desugar.cc
is the only one that wants to flip RewriterSynthesized off, and we
probably that number to never grow vs the number of rewriter passes.