Make ast::MK::Method assumes RewriterSynthesized (#2672)
* More accurate return type for ast::MK::Method0

* ast::MK::Method assumes RewriterSynthesized (common case)

* The two places where the new default isn't right

* Update callsites to stop passing flags

* But actually, this can be synthesized

I tested this on pay-server, and this one which was not being marked as
RewriterSynthesized is actually safe to mark as such, and probably
should be for consistency. That means that the call site in Desugar.cc
is the only one that wants to flip RewriterSynthesized off, and we
probably that number to never grow vs the number of rewriter passes.

* MK::Method* -> MK::SyntheticMethod* (#2674)

* MK::Method* -> MK::SyntheticMethod*

* rg -l 'MK::Method(\b|0|1)\b'  | xargs sed -i -e 's/MK::Method/MK::SyntheticMethod/g'

* tools/scripts/format_cxx.sh

* Fix rename not caught by regex

* Implement ast::MK::SyntheticMethod in terms of ast::MK::Method

This makes Desugar.cc a little more straightforward (don't have to unset
isRewriterSynthesized, which is less surprising to a reader).