Don't check sig of `Decorator#initialize`
`Decorator.new` is being called as part of the initialization of `sorbet-runtime` library, which in turn load the signature for `Decorator#initialize` and sets the internal `@has_read_default_checked_level` to `true` before the library consumer even has a chance to set it. Thus, setting  `T::Configuration.default_checked_level` always fails after doing a `require 'sorbet-runtime'`.

By changing the signature on `Decorator#initialize` to be never checked, we still protect the static checks on the method but make sure the runtime library does not get initialized too early.