Report per-edit (from the client) diagnostic latency (#2749)
Creates a more representative latency metric for Sorbet. For every edit (even if combined!), this will report how long it takes to provide a diagnostics update containing that edit.

The start time for this metric is when the editor sends an edit to Sorbet. The end time is when Sorbet completes typechecking the combined LSPFileUpdates containing that edit.