Simplify LSP cancelation decision.
If a new edit comes in and it will take the slow path relative to the last committed update, it should cancel the currently running slow path.

There's no need to check if it would take the fast path with the currently running slow path, since there exists no update such that it takes the fast path relative to the last edit AND takes the fast path with the currently running slow path.