Fix method parameter serialization for delegation in Ruby 2.7.
Technically, this should probably generate the method signature as
`foo(...)`, but Sorbet doesn't know how to parse that properly. Instead,
it replaces it with `foo(*args, &blk)`.

This is mostly just a workaround to make the hidden.rbi generator work.
I'm not 100% sure what happens if  you actually use a method with these
type signatures.