LSP: Fix segfaults in hashing. (#2787)
* Fix #1: Don't flush trees to disk unless global state was written.

There is an edge case ("critical" errors) where we might write trees but not the corresponding global state.

* Fix #2: Read everything in a single transaction, write everything in a single transaction.

Previously, we might read GlobalState and cached trees in different transactions. If two Sorbet processes interleaved file reads and writes in just the right way, a process may find itself with an old nametable with new trees referring to names not on the old nametable -- which then gets committed to disk before segfaulting in namer!

I have also consolidated logic related to creating KVStore and caching gs/trees into a new package (//main/cache). I plan to move caching routines from pipeline into this new package soon.

There are enforces in debug builds to maintain this invariant.

* Fix #3: Don't write cache if it has been mutated since the last read.

Previously, it was theoretically possible for the following interleaving to occur:

* Sorbet process 1 reads GS and trees, and gives up DB lock.
* Sorbet process 2 writes GS and adds a new tree to the DB that references new names.
* Sorbet process 1 overwrites process 2's GS.

Now, in subsequent sessions, Sorbet might try to use the tree from process 2, which references names that are not in process 1's GlobalState -- causing segfaults.

I have fixed this issue by adding a random integer to GlobalStates before writing it to the database. If it changes between the read and write transactions, Sorbet will not write its changes to disk.