lldbinit: Add llvmsourcemap command
This only works on Linux right now, because I couldn't figure out a
good, canonical symbol to list. When debugging Ruby, we can pretty much
always `list main`, but since LLVM is a library, we need to find some
other canonical symbol that we can list. All the ones I tried seemed
aggressively ad-hoc and/or had a bunch of overloads so the symbol
existed at 12 different locations.

On macOS, the path to llvm looked something like this:

    (lldb) list CreateCall
    File: /private/var/tmp/_bazel_jez/fd5fca88654f2ca281df28e218fddc0e/execroot/com_stripe_sorbet_llvm/external/llvm/include/llvm/IR/IRBuilder.h
    File: /private/var/tmp/_bazel_jez/fd5fca88654f2ca281df28e218fddc0e/execroot/com_stripe_sorbet_llvm/external/llvm/include/llvm/IR/IRBuilder.h
    File: /private/var/tmp/_bazel_jez/fd5fca88654f2ca281df28e218fddc0e/execroot/com_stripe_sorbet_llvm/external/llvm/include/llvm/IR/IRBuilder.h
    ...