Ignore `MultipleBehaviourDefs` error from suggest-typed
`MultipleBehaviourDefs` error seems to be an internal Stripe-only error
that is raised for when a type has behaviour coming from multiple files.

Under normal circumstances, that error is never visible for end-users
unless `--stripe-mode` flag is enabled. However, the error was still
being considered for setting the minimum error level of each file, and
was thus breaking the behaviour of `--suggest-typed`. As a result many
files that were passing type checking with strictness levels `false` or
above were being suggested `ignore` instead.

This commit fixes the problem by ignoring `MultipleBehaviorDefs` when
setting the minimum error level of a file.