Inlcude IREmitterHelpers.h at full relative path
This uncovered some bugs, where Payload.h was only being included in one
place, and that one place had always happened to be after
IREmitterHelpers.h. That's not a good look, so I added some forward
declares and #include's to Payload.h so that it can work.

I also happened to see that we were using some `sorbet::` qualifiers
where we didn't need them, so I dropped them for brevity.