funcId -> rubyBlockId
There's no such thing as a funcId except in this one local variable.
Maybe the intention was that we wanted to keep track of some notion of
the ID of an llvm::Function, but since it's already one-to-one with
rubyBlockId, I think naming it this way makes it more clear (rubyBlockId
is something that already exists and has meaning elsewhere in the
codebase, and funcId is essentially equal to it).