Remove <chrono> from Counters
After changing the clock and changing the vectors, we found a lot of the
remaining slowness was overhead of using the chrono library to compute
on durations. Rather than figure out how to make our usages of the
chrono library faster, we decided we'd forgoe the typesafety of their
abstractions and just multiply or divide ints to get the same behavior.

Again, we observed that this made Timers faster via `perf`.