Make Timers faster (#2831)
* Replace chrono::steady_clock::now() with clock_gettime(CLOCK_MONOTONIC_COARSE, &tp)

We looked at this benchmark[0] to see how fast various C++ clocks were
(for the purpose of selecting CLOCK_MONOTONIC_COARSE in the first place)
and then used `perf` to verify that it made timers faster.

[0]: https://stackoverflow.com/questions/48609413/fastest-way-to-get-a-timestamp

* Make vectors into pointers, so we don't have to pay to destruct them

We found that we were paying a lot just to construct and destruct
vectors in Timers that were usually empty. We thought about making it
into an InlinedVector<..., 1>, but decided that we'd rather make it into
a unique_ptr because it would be even simpler.

This also had the side-effect of catching various places where we could
have been using `move(...)` to avoid copying when calling a helper
function (i.e., using `unique_ptr` made the moves required, whereas
before they would have been an optional perf improvement).

Again, we verified this by using `perf` to look at where timers were
spending their time.

* Remove <chrono> from Counters

After changing the clock and changing the vectors, we found a lot of the
remaining slowness was overhead of using the chrono library to compute
on durations. Rather than figure out how to make our usages of the
chrono library faster, we decided we'd forgoe the typesafety of their
abstractions and just multiply or divide ints to get the same behavior.

Again, we observed that this made Timers faster via `perf`.

* Add more comments

* Fix forward declare

* Bump threshold

* Fix LSP testing

* Revert everything having to do with unique_ptr

* Don't use microsecond pause times to differentiate timers.

We use a coarse-granularity timer now.

* Apparently coarse timestamps are really coarse.

* Revert "Revert everything having to do with unique_ptr"

This reverts commit 4688883ab7ba595cbc0c74f7acbade0d71e53847.

* Ignore one more comparison

* Stop doing stupid stuff with raw pointers

Co-authored-by: John Vilk <jvilk@stripe.com>