Web trace file format requires floating point timestamps
Also, I had forgotten to change "now" to use clock_gettime_coarse.
It was instead using steady_clock still, which generates completion
incomparable timestamps (so when those timestamps were both converted to
unit-less floating point numbers, it looked like the counters had a
timestamp of weeks in the future compared to the timestamps of the
floating point numbers).