Add print statement to help lldb debugging (#2850)
There used to be something in the bazel test output that would show the
command and arguments it ran when running a test. This was super
helpful, because it could be copy/pasted after `lldb --` and then you
could start debugging.

I think probably one of the bazel upgrades killed that. Instead, I opted
to just add back the print statement manually in our forwarder:

    ==================== Test output for //test:test_LSPTests/testdata/lsp/completion/constants_all_kinds:
    bazel-bin/test/test_corpus_runner --single_test=test/testdata/lsp/completion/constants_all_kinds.rb --gtest_filter=LSPTests/\*
    Note: Google Test filter = LSPTests/*
    [==========] Running 1 test from 1 test suite.
    ...