Add support for tracing tests
This helped me diagnose where the slowness in the constant_all_kinds
completion test was coming from, so I figure we may as well leave it
checked in for ease of future debugging.

For future reference, my preferred way of using this flag is:

1.  Run the test via bazel, to ensure that all the files are built:

    `bazel test --config=dbg --test_output=all //test:test_LSPTests/testdata/my/test`

2.  In the output near the top, there will be a line that shows how the
    test runner was invoked (with what flags):

    `bazel-bin/test/test_corpus_runner --single_test=...  --gtest_filter=...`

3.  I copy / paste that line into my terminal, and append a `--web-trace-file=trace.json` argument.

That drops a `trace.json` file into the top-level of the Sorbet folder.

(This "run the test directly at the command line" technique is also
useful, because you can take the binary and run `perf` or `lldb` over it
too, or load it up in `Instruments.app`).