Emit raw, pre-formatted values for timestamps (#2858)
This gets us back to parity with what tracing.cc was doing before I
refactored it to use rapidjson instead of fmt::format_to.

We use `fmt::format` to turn a fixed `int64_t` into a floating point
number with 3 digits of precision (rather than dividing by 1000.0).
`writer.String` takes that std::string and treats it as a JSON string
when serializing it (so the type of `"ts"` is a JSON strong).  On the
other hand `writer.RawValue` takes that std::string and treats it as an
already-formatted JSON value, with a given type. That makes the type for
`"ts"` back into a number.