Second param of `Kernel.raise` can be anything
While Ruby documentation claims that the second parameter
is expected to be a `message` which is a `String` object,
it allows users to pass in any object for that parameter.
The `message` parameter is passed to the given exception class'
constructor directly.

Many libraries and frameworks already use the pattern of passing
in arbitrary objects for exceptions via the `message` parameter
already like:

```
class ClientError < StandardError
  attr_reader :http_response

  def initialize(http_response)
    @http_response = http_response
    super("some dynamic message")
  end
end

raise ClientError, http_response
```