Fix splat destructuring type information
The desugaring for processing splat destructuring uses `.slice` which
returns a nilable result, thus corrupting the type of the splat
parameter which should always be a non-nilable array type.

The quick fix is to add a `T.must` around that `.slice` call to fix the
semantics of the desugaring, which is what this commit does.