release-debug-*: Replace --config=dbg with --config=debugsymbols
--config=dbg includes `-O0`, which is great for using with a debugger,
but is terribly slow on Stripe's codebase:

- before this change: ~25 minutes for one `release-debug-linux` type
  check run
- after this change: ~2 minutes

release-common already includes `--config=backtracesymbols`, but this
change explicitly overrides that to have `--config=debugsymbols`. This
will have things like names of local variables in the frame.

That being said, because so many things have been inlined, optimized,
and LTO'd, it's possible that certain methods you'd like to be able to
call on those local variables might not exist, so it's not perfect for
debugging.

Ultimately, the goal here is to have `release-debug-linux` be something
which runs fast and runs all ENFORCE checks.