Make the canonical loc of a symbol the strictest (#2873)
* Make the canonical loc of a symbol the strictest

* Add failing test

* Change some sealed locs

After changing the canonical loc, some symbols' locs moved around. In
particular, if you had

```
class TrueClass
end
```

in a Ruby file, previously the loc of the Symbol for `TrueClass` would
be this Ruby file, but now it is always the one in the stdlib, unless
the current file's strictness level is greater than or equal to the RBI
(which in the case of TrueClass is __STDLIB_INTERNAL).

Instead, for checking sealed, we now explicitly check the file where the
`include` was written.

Then the problem becomes: what if that RBI file definition isn't in the
stdlib but instead generated by reflection? Then then it looks like the
RBI file isn't allowed to include some constant.

So instead, I've opted to recreate the way that the "canonical loc"
selection used to work with a helper directly in validator.cc, to
preserve the old behavior.

* RBIs that don't actually write sigs for all their methods

* Add a representative test

* Reformat