Narrower types in ast::MK helpers
This is nice because it means that calling one of these helpers is no
worse and having created the object yourself. Specifically, the caller
will be able to access fields of the object that was just created, like
`flags` in the case of `ast::Send`.

The changes from `x.swap(y)` to `x = move(y)` are all harmless, because
in all cases were were constructing temporary values like

    unique_ptr<Expression> result;

and swapping into them. `swap` only allows swapping when the objects are
the exact same. For example, you can't swap an Expression into a Send.
But you can `move` a `Send` into an `Expression`, and since we never
cared about what was left in the non-result variable, it's fine to do
this.