Report incorrect computed_by usage directly
I'm trying to split processProp into two halves: the half that parses a
prop into it's data, and the half that consumes that data to write out
new expressions.

The logic for computed_by violates that: it writes out expressions and
parses the data at the same time, so I rewrote it to not do that.

I opted to introduce a new error, rather than relying on the error
assertion from a `T.let` in infer. There's two reasons:

1. Don't have to prematurely change the tree before we've parsed a prop
2. The error is more specific: not only does it have to have type
   Symbol, it has to syntactically be a Symbol literal.