Split processProp into two functions
At this point, parseProp type checks, but doesn't do much useful (drops
most of what it parsed on the floor) and the new processProp doesn't
even typecheck. But I wanted to start somewhere to show a small diff.

The contract here is:

- parseProp doesn't mutate the prop it's parsing at all, so that we can
  early exit at any point and not have consumed any part of the tree
- eventually processProp will be a pure function of the PropInfo that
  parseProp created.

To make this work out, rather than worrying about where we're copy and
where we're not, I unconditionally copy the entire `rules` hash. Some
parts of parsing the prop will consume this parsed `rules` hash, but
since it's a copy, it won't affect the underlying prop.