Default initialize NameRef to _id == 0
FileRef and SymbolRef both initialize to `_id == 0`, which is a valid
FileRef or SymbolRef for which `.exists()` is `false`.

Before this change, NameRef initialized `_id` to `-1`, which was an
invalid NameRef, which was never checked in any internal consistency
checks, and meant that `.exists()` was `true`. This lead to surprising
behaviors and errors far from where the problem needed to be corrected.

In the sake of consistency and convenience, we now default initialize
`NameRef` to `_id(0)`.