Get rid of warning on every bazel invocation (#2923)
I would see this message on every bazel invocation:

    WARNING: Duplicate rc file: /Users/jez/stripe/sorbet/.bazelrc.local is read multiple times, most recently imported from /private/var/folders/5p/z_n7b95s1yz02fdjwvk8cwp80000gn/T/bazel_bazel_rc_1587612772

It's because bazel reads **all** config files provided to it, as they
have in the docs:

    https://docs.bazel.build/versions/master/guide.html#bazelrc

My guess is that there was a point in time when this wasn't the case at
Stripe, and we needed these lines. But as far as I can tell, we no
longer need them, and it meant that the `try-import` in our `.bazelrc`
in this repo was getting processed twice and printing that warning.