Implement non_forcing_is_a? in sorbet-runtime
See the tests for now; docs will follow.

It uses `autoload?` to guard `const_defined?` in a loop to attempt to
resolve a constant. The idea is that if a constant is not yet loaded,
the `is_a?` check is necessarily `false`, and you don't need to force
the constant to load to arrive at that result. This makes it possible to
break up certain kinds of load-time / runtime dependencies.

This feature is backed by Sorbet statically, so even though the runtime
won't know whether a constant actually exists at runtime, Sorbet will be
able to step in and guarantee that there were no typos in the name. In
particular, the runtime returns `false` if both the constant doesn't
exist and if it might exist but we can't tell because it's not yet
loaded. Another option would have been to raise in cases when we were
sure that a constant didn't exist (because there wasn't even an autoload
registered for it), but I figured that would be more surprising
(makes it more load-order dependent).

I think at Stripe there are also some use cases to have a variant of
this that's backed by `Module#<` as well, for comparing class objects,
not instances of classes. We should feel free to implement that when the
need arrises.