Remove useless same-loc test
This test was originally introduced to say that reporting errors came in
some consistent order based on Loc.  It was added 18 months ago.

In the time since, we've introduced a bunch more tests that make this
assumption more explicit (LSP tests with `^^^` assertions) and made the
pipeline itself more resilient to this behavior.

On the other hand, I frequently end up recording this test for
completely unrelated changes, to the point where the actual snapshot is
meaningless at this point. So I'm voting we remove it.