Partially revert the last change
Turns out we have a bunch of code in cfg/builder/ that makes a
distinction between "this field has never been set" vs "this field has
been set but it's been set to a meaningless name"

We could solve this one of two ways:

- change how cfg/builder/ uses names
- change how names behave

I choose the second approax, which is basically keeping the status quo,
almost as if this change hadn't been written.

The one change that I've made is that I've made `.exists()` return false
for both when the nameref is uninitialized and when it's initialized to
noName, which I think is less surprising.