Simplify completion
The code essentially tried to say "if T::Sig::WithoutRuntime.sig then
try to suggest a sig, but always abort halfway through"

That used to be required, because we re-wrote all `sig` calls at the
top-level to T::Sig::WithoutRuntime, even if they didn't have the right
args. Now, if they don't have the right number of args, we leave the
original call to `sig` in the tree, which means that the only case we'd
get to trySuggestSig with a `T::Sig::WithoutRuntime.sig` completion item
is if the user literally wrote that out and asked for a completion.
That's far more rare, and what we were trying to avoid supporting in the
first place by short circuiting, so now the code is a lot more succinct.

This is kind of confusing explained with words, so I wrote some tests.